Testing improvements * Docker copies case-INsensitive test repo to a case-sensitive one, * -i option: take_test.py leaves output in place on success for interactive browsing via take_test.sh in docker.

+++ tests/take_test.py
@@ -61,6 +61,14 @@ parser.add_argument(
     default=None
 )
 
+parser.add_argument(
+    '-i',
+    action='store_true',
+    help=('Keep output (for browsing) even when all tests successful.'),
+    default=None
+)
+
+
 args = parser.parse_args()
 candidate_spec_file = args.candidate_spec_file
 domain = args.domain
@@ -71,6 +79,8 @@ test_spec = args.test_spec
 if test_spec is not None:
     if type(test_spec) is not list:
         test_spec = [test_spec]
+
+interactive = args.i
 
 
 # #######################################################
@@ -237,10 +247,13 @@ log.debug('')
 # Tear down if success
 log.info('=================================================================')
 if pytest_return == 0:
-    log.info('All tests successful: tear down test.')
-    log.debug('')
-    shutil.rmtree(candidate_spec['repos_dir'])
-    shutil.rmtree(candidate_spec['test_dir'])
+    if interactive:
+        log.info('All tests successful but -i leaves files for interactive browsing.')
+    else: 
+        log.info('All tests successful: tear down test.')
+        log.debug('')
+        shutil.rmtree(candidate_spec['repos_dir'])
+        shutil.rmtree(candidate_spec['test_dir'])
     log.debug('')
 else:
     if pathlib.PosixPath(candidate_spec['repos_dir']).exists():
@@ -248,7 +261,7 @@ else:
         log.info('Repos: ' + candidate_spec['repos_dir'])
     else:
         log.info('Some tests failed: leaving tests in:')
-        
+
     log.info('Tests: ' + candidate_spec['test_dir'])
     log.debug('')
 
@@ -290,3 +303,6 @@ all_specs = { 'Candidate': candidate_spec,
 
 for key, value in all_specs.items():
     log_spec(value, key)
+
+
+sys.exit(pytest_return)
